# docs/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/items/{id}/publishing-space": {
      "get": {
        "operationId": "PublishingDraftsController_space",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channelId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PublishingDrafts"
        ]
      }
    },
    "/api/items/{id}/publishing-draft": {
      "post": {
        "operationId": "PublishingDraftsController_save",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "PublishingDrafts"
        ]
      }
    },
    "/api/packages/{id}/usable": {
      "get": {
        "operationId": "PublishingDraftsController_usable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PublishingDrafts"
        ]
      }
    },
...
    "/api/collections/{id}": {
      "get": {
        "operationId": "CollectionsController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Collections"
        ]
      }
    },
    "/api/collections/{id}/download": {
      "get": {
        "operationId": "CollectionsController_download",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Collections"
        ]
      }
    },
...
  },
  "info": {
    "title": "ToMeBoutique Workbench API",
    "description": "",
    "version": "1.0.0-rc.14",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tome_session"
      }
    },
...
  }
}